refactor: dry cluster definitions and move helper code in own file
